+++ front/react_app/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Cycle from "./pages/Cycle";
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 import Trash from "./pages/Trash";
 import "./App.css";
